Stan is bad at type casting

+++ Tests/UnitTests.cs
@@ -83,7 +83,7 @@ namespace TravisCILab
         {
             Assert.AreEqual(2, Program.Divide("2", "1"));
             Assert.AreEqual(1.5, Program.Divide("3", "2"));
-            Assert.AreEqual(5/7, Program.Divide("5", "7"));
+            Assert.AreEqual((double)5/(double)7, Program.Divide("5", "7"));
         }
 
         [Test]
